requirements.txt
gtts
playsound
openai
python-dotenv
speech_recognition